+++ spec/opensearch/api/actions/security/patch_audit_configuration_spec.rb
@@ -1,0 +1,40 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+
+# This code was generated from OpenSearch API Spec.
+# Update the code generation logic instead of modifying this file directly.
+
+# frozen_string_literal: true
+
+require_relative '../../../../spec_helper'
+
+describe 'client.security#patch_audit_configuration' do
+  let(:expected_args) do
+    [
+      'PATCH',
+      '_plugins/_security/api/audit',
+      {},
+      {},
+      {}
+    ]
+  end
+
+  let(:client) do
+    Class.new { include OpenSearch::API }.new
+  end
+
+  it 'requires the :body argument' do
+    expect do
+      client.security.patch_audit_configuration
+    end.to raise_exception(ArgumentError)
+  end
+
+  it 'performs the request with all optional params' do
+    expect(client_double.security.patch_audit_configuration(
+      body: {}
+    )).to eq({})
+  end
+end